Add cut and delete previous clip feature
Cut the clip and removes the previous one and the all space just before the
the removed item. The timeline will be moved at the start of the clip
that has been kept.

Cutting an empty space will just empty the space.

Undo will not handle all steps at once. Several will be needed.

Signed-off-by: Vinicius Tinti <viniciustinti@gmail.com>